requirements.txt
wheel
flask
flask-compress
pymongo
mysql
cryptography
pyperclip
python-dotenv
passlib
psutil
bcrypt
gunicorn
requests
PySimpleGUI
flask_socketio
pypiwin32
